Added the web page (front end)
Pretty much no js work done. Tried to get fav buttons to create a popup but it hasn't worked out properly for whatever reason. The popup is not styled yet. AI button functions haven't been started at all but it should just be as easy as creating another popup or a dropdown or something that contains the text. Remember that this isn't the final product so we can afford to not have a few features fully complete.